requirements.txt
elevenlabs>=1.0.0
av>=10.0.0
Pillow>=10.0.0
numpy>=1.24.0
moviepy>=1.0.0
openai-whisper>=20231117
python-dotenv>=1.0.0
mcp>=1.0.0
google-genai>=0.3.0
rembg>=2.0.0
